ASCII done!
I'm pretty proud that I finally came up with a nice high level solution for
this. From the beginning, I was thinking in terms of having just one big
list of Chars using width and height of character for outsetting in the
big list. Which definitely could be done in Haskell, but felt wrong.

Nevertheless, this feels nice. And I think it's readable enough.